backend/requirements.txt
# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3

# Google Sheets API
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Environment variables
python-dotenv==1.0.1

# CORS
python-multipart==0.0.6
